# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for BancoPeriodosFixosModo.
const (
	DaApi  BancoPeriodosFixosModo = "da-api"
	DoHtml BancoPeriodosFixosModo = "do-html"
	Lista  BancoPeriodosFixosModo = "lista"
)

// Valid indicates whether the value is a known member of the BancoPeriodosFixosModo enum.
func (e BancoPeriodosFixosModo) Valid() bool {
	switch e {
	case DaApi:
		return true
	case DoHtml:
		return true
	case Lista:
		return true
	default:
		return false
	}
}

// Defines values for OfertaFiabilidade.
const (
	Confirmada   OfertaFiabilidade = "confirmada"
	EmDuvida     OfertaFiabilidade = "em_duvida"
	PorConfirmar OfertaFiabilidade = "por_confirmar"
)

// Valid indicates whether the value is a known member of the OfertaFiabilidade enum.
func (e OfertaFiabilidade) Valid() bool {
	switch e {
	case Confirmada:
		return true
	case EmDuvida:
		return true
	case PorConfirmar:
		return true
	default:
		return false
	}
}

// Defines values for PedidoEuriborIndexante.
const (
	N12m PedidoEuriborIndexante = "12m"
	N3m  PedidoEuriborIndexante = "3m"
	N6m  PedidoEuriborIndexante = "6m"
)

// Valid indicates whether the value is a known member of the PedidoEuriborIndexante enum.
func (e PedidoEuriborIndexante) Valid() bool {
	switch e {
	case N12m:
		return true
	case N3m:
		return true
	case N6m:
		return true
	default:
		return false
	}
}

// Defines values for PedidoFinalidade.
const (
	Arrendamento PedidoFinalidade = "arrendamento"
	Propria      PedidoFinalidade = "propria"
	Secundaria   PedidoFinalidade = "secundaria"
)

// Valid indicates whether the value is a known member of the PedidoFinalidade enum.
func (e PedidoFinalidade) Valid() bool {
	switch e {
	case Arrendamento:
		return true
	case Propria:
		return true
	case Secundaria:
		return true
	default:
		return false
	}
}

// Defines values for PedidoRateType.
const (
	Fixa     PedidoRateType = "fixa"
	Mista    PedidoRateType = "mista"
	Variavel PedidoRateType = "variavel"
)

// Valid indicates whether the value is a known member of the PedidoRateType enum.
func (e PedidoRateType) Valid() bool {
	switch e {
	case Fixa:
		return true
	case Mista:
		return true
	case Variavel:
		return true
	default:
		return false
	}
}

// Banco defines model for Banco.
type Banco struct {
	// EuriborImposto Quando não-nulo, o banco impõe este indexante.
	EuriborImposto *string `json:"euribor_imposto"`

	// EuriborOpcoes Vazio não é «não sei» — é «o banco impõe o seu e ignora a escolha».
	//
	// Example: ["3m","6m","12m"]
	EuriborOpcoes []string `json:"euribor_opcoes"`

	// Id Example: novobanco
	Id             string       `json:"id"`
	IdadeMaximaFim int          `json:"idade_maxima_fim"`
	Inputs         []BancoInput `json:"inputs"`

	// Nome Example: Novo Banco
	Nome              string                 `json:"nome"`
	Notas             []string               `json:"notas"`
	PeriodosFixos     []int                  `json:"periodos_fixos"`
	PeriodosFixosModo BancoPeriodosFixosModo `json:"periodos_fixos_modo"`
	PrazoMax          int                    `json:"prazo_max"`
	PrazoMin          int                    `json:"prazo_min"`
	Produtos          []Produto              `json:"produtos"`
}

// BancoPeriodosFixosModo defines model for Banco.PeriodosFixosModo.
type BancoPeriodosFixosModo string

// BancoInput defines model for BancoInput.
type BancoInput struct {
	// Chave Example: valor_imovel
	Chave string `json:"chave"`

	// Nota Presente quando ajuda a explicar porque o banco não usa o campo.
	Nota *string `json:"nota,omitempty"`
	Usa  bool    `json:"usa"`
}

// BancosResposta defines model for BancosResposta.
type BancosResposta struct {
	Bancos          []Banco         `json:"bancos"`
	InputsCanonicos []InputCanonico `json:"inputs_canonicos"`
}

// DetalheErro defines model for DetalheErro.
type DetalheErro struct {
	// Campo O campo que falhou, quando aplicável.
	Campo *string `json:"campo,omitempty"`

	// Codigo Example: pedido_invalido
	Codigo   string `json:"codigo"`
	Mensagem string `json:"mensagem"`
}

// Fase defines model for Fase.
type Fase struct {
	AteMes    int     `json:"ate_mes"`
	Prestacao float64 `json:"prestacao"`
	Taxa      float64 `json:"taxa"`
}

// InputCanonico defines model for InputCanonico.
type InputCanonico struct {
	// Chave Example: valor_imovel
	Chave string `json:"chave"`

	// Rotulo Example: Valor do imóvel
	Rotulo string `json:"rotulo"`

	// Tipo Example: dinheiro
	Tipo string `json:"tipo"`
}

// Oferta ⚠️ Quando `aplicado` não está vazio, os números não correspondem ao pedido — a app é obrigada a mostrar a nota junto do valor. `erro` só aparece quando `sucesso` é falso.
type Oferta struct {
	// Aplicado O que o banco usou de facto, quando difere do pedido. Vazio = igual ao pedido.
	Aplicado  *map[string]interface{} `json:"aplicado,omitempty"`
	BancoId   string                  `json:"banco_id"`
	BancoNome string                  `json:"banco_nome"`

	// CapturadoEm Quando o preço foi MEDIDO no banco, durante o varrimento — não quando esta resposta foi calculada (isso é o `calculado_em` da Comparacao).
	// ⚠️ Presente sempre que `sucesso` é verdadeiro, e é o campo mais importante desta lista para a honestidade do produto: é o único que diz à pessoa que está a ver um preço de ontem à noite e não de agora. A app é obrigada a mostrá-lo. Escondê-lo apresentaria dados varridos como se fossem uma consulta ao vivo ao banco, que é exactamente o que este serviço não faz.
	CapturadoEm *time.Time `json:"capturado_em,omitempty"`

	// EmCache Verdadeiro quando esta oferta se serviu da cache (§7.6) em vez de se perguntar ao banco agora.
	// ⚠️ **Não é o mesmo que `capturado_em`, e nenhum dos dois substitui o outro.** O `capturado_em` diz de QUANDO é o preço, e num acerto de cache é o instante em que se falou com o banco — nunca o de agora. Este diz se ESTE pedido chegou a sair para o banco. Uma oferta fresca e um acerto de um segundo atrás têm `capturado_em` quase igual e `em_cache` diferente.
	// ⚠️ Ausente vale `false`, e não leva `default:` de propósito — pela mesma razão do `fiabilidade` acima: o `openapi-typescript` traduz um campo com omissão para um campo OBRIGATÓRIO no tipo gerado.
	EmCache *bool `json:"em_cache,omitempty"`

	// Erro Estruturado de propósito: `codigo` é para a app decidir, `mensagem` é para a pessoa ler, em português.
	Erro             *OfertaErro `json:"erro,omitempty"`
	EuriborIndexante *string     `json:"euribor_indexante,omitempty"`
	EuriborValor     *float64    `json:"euribor_valor,omitempty"`
	Fases            *[]Fase     `json:"fases,omitempty"`

	// Fiabilidade O que se sabe sobre a grelha de onde este preço saiu, e não sobre o preço em si (KAN-49). A sonda confirma — com ~4 pedidos por banco — se a grelha guardada ainda descreve o banco.
	// ⚠️ `em_duvida` quer dizer que a última sondagem DISCORDOU da grelha e que ainda não houve varrimento que resolvesse a discordância. A app é **obrigada** a mostrar a nota que vem em `notas`, junto do número — é a mesma regra do `aplicado`.
	// ⚠️ `confirmada` e `por_confirmar` mostram-se com **silêncio**. Uma marca de «confirmada» em toda a gente é ruído com aspecto de informação, e treina quem lê a saltar a única que importa.
	// ⚠️ **Ausente vale `por_confirmar`**, e nunca `confirmada`: uma omissão que valesse «confirmada» afirmava sobre a série inteira uma coisa que ninguém mediu. Este servidor manda-o sempre; o campo fica opcional porque `required` só tem o que uma oferta não pode não ter.
	// ⚠️ E **não leva `default:`** de propósito. Levava, e o `openapi-typescript` traduz um campo com omissão para um campo OBRIGATÓRIO no tipo gerado — a app passava a afirmar que a resposta traz sempre este campo, que é precisamente o que o `required` diz que não. A regra fica na descrição, onde não mente ao gerador.
	Fiabilidade *OfertaFiabilidade `json:"fiabilidade,omitempty"`

	// Mtic ⚠️ DERIVADO, pela mesma razão e sobre o mesmo modelo de encargos que a `taeg`. Ver `pressupostos`.
	Mtic  *float64  `json:"mtic,omitempty"`
	Notas *[]string `json:"notas,omitempty"`

	// Pressupostos As hipóteses sob as quais a `taeg` e o `mtic` desta oferta foram derivados, em português e legíveis por uma pessoa.
	// ⚠️ Lista à parte de `notas`, e não misturada nela, de propósito: as duas têm estatutos diferentes. Uma `nota` é um aviso sobre o que aconteceu a ESTE pedido — o banco encurtou o prazo, o degrau de LTV não estava resolvido. Um pressuposto é uma hipótese de cálculo que a MCD obriga a declarar junto do número que dela depende (Anexo I, Parte II; Anexo II). Empacotadas na mesma lista, a app ficava sem forma de as apresentar como o que são.
	// ⚠️ Não vazia sempre que `taeg` ou `mtic` vêm preenchidos. Vazia com um deles preenchido é defeito nosso, e não um caso legítimo.
	Pressupostos *[]string `json:"pressupostos,omitempty"`

	// PrestacaoMensal Amortização francesa sobre a `tan` e o montante pedido. É aritmética exacta, não medição — e é por isso que responde ao montante desta pessoa, e não ao do cenário de referência.
	PrestacaoMensal   *float64  `json:"prestacao_mensal,omitempty"`
	ProdutosAplicados *[]string `json:"produtos_aplicados,omitempty"`

	// Spread ⚠️ Do INTERVALO de LTV medido que contém o rácio pedido, e não de uma banda assumida. Num degrau que não se conseguiu resolver é o lado mais CARO do intervalo (Directiva 2014/17/UE, Anexo I, Parte II, alínea (d)), e `notas` traz a frase que nomeia os dois lados.
	Spread  *float64 `json:"spread,omitempty"`
	Sucesso bool     `json:"sucesso"`

	// Taeg ⚠️ DERIVADA, não medida. Ver `pressupostos`, que é obrigatório sempre que este campo vem preenchido.
	// A TAEG depende dos encargos (comissões, imposto, seguros) e o seguro de vida depende de quem pede — e a série de mercado é varrida com um titular NEUTRO e fictício, sobre um cenário de referência fixo (§4). Não existe, portanto, TAEG medida para esta pessoa. O que se faz é atribuir a encargos a diferença entre a TAEG e a TAN observadas no ponto de referência, e reamortizar sobre os fluxos deste pedido.
	// ⚠️ A repartição desses encargos entre comissões, imposto e seguro NÃO está medida: é inferida. Quem lê este número tem de o tratar como indicativo, e a app é obrigada a mostrar `pressupostos` junto dele — é o que o Anexo I e o Anexo II da MCD mandam fazer a um valor que depende de hipóteses declaradas.
	Taeg *float64 `json:"taeg,omitempty"`

	// Tan Spread do degrau de LTV que contém o rácio pedido, mais a Euribor do tenor aplicável — ambos medidos. Na fixa e na mista, a taxa da fase fixa é a observação do período que o banco pratica.
	Tan *float64 `json:"tan,omitempty"`
}

// OfertaFiabilidade O que se sabe sobre a grelha de onde este preço saiu, e não sobre o preço em si (KAN-49). A sonda confirma — com ~4 pedidos por banco — se a grelha guardada ainda descreve o banco.
// ⚠️ `em_duvida` quer dizer que a última sondagem DISCORDOU da grelha e que ainda não houve varrimento que resolvesse a discordância. A app é **obrigada** a mostrar a nota que vem em `notas`, junto do número — é a mesma regra do `aplicado`.
// ⚠️ `confirmada` e `por_confirmar` mostram-se com **silêncio**. Uma marca de «confirmada» em toda a gente é ruído com aspecto de informação, e treina quem lê a saltar a única que importa.
// ⚠️ **Ausente vale `por_confirmar`**, e nunca `confirmada`: uma omissão que valesse «confirmada» afirmava sobre a série inteira uma coisa que ninguém mediu. Este servidor manda-o sempre; o campo fica opcional porque `required` só tem o que uma oferta não pode não ter.
// ⚠️ E **não leva `default:`** de propósito. Levava, e o `openapi-typescript` traduz um campo com omissão para um campo OBRIGATÓRIO no tipo gerado — a app passava a afirmar que a resposta traz sempre este campo, que é precisamente o que o `required` diz que não. A regra fica na descrição, onde não mente ao gerador.
type OfertaFiabilidade string

// OfertaErro Estruturado de propósito: `codigo` é para a app decidir, `mensagem` é para a pessoa ler, em português.
type OfertaErro struct {
	// Codigo Example: prazo_impossivel
	Codigo   string `json:"codigo"`
	Mensagem string `json:"mensagem"`
}

// OfertaPedido O pedido para UM banco. Não leva lista de bancos: o banco vem no caminho, e os produtos são só os desse banco.
type OfertaPedido struct {
	Pedido Pedido `json:"pedido"`

	// Produtos Ids dos produtos escolhidos deste banco.
	Produtos *[]string `json:"produtos,omitempty"`
}

// Pedido defines model for Pedido.
type Pedido struct {
	EuriborIndexante *PedidoEuriborIndexante `json:"euribor_indexante,omitempty"`
	Finalidade       PedidoFinalidade        `json:"finalidade"`

	// FixedPeriodYears Anos de período fixo; aplica-se a `fixa` e `mista`.
	FixedPeriodYears *int  `json:"fixed_period_years,omitempty"`
	GarantiaPublica  *bool `json:"garantia_publica,omitempty"`
	JaCliente        *bool `json:"ja_cliente,omitempty"`

	// Localizacao Example: continente
	Localizacao string         `json:"localizacao"`
	Montante    float64        `json:"montante"`
	PrazoAnos   int            `json:"prazo_anos"`
	RateType    PedidoRateType `json:"rate_type"`
	Titulares   []Titular      `json:"titulares"`
	ValorImovel float64        `json:"valor_imovel"`
}

// PedidoEuriborIndexante defines model for Pedido.EuriborIndexante.
type PedidoEuriborIndexante string

// PedidoFinalidade defines model for Pedido.Finalidade.
type PedidoFinalidade string

// PedidoRateType defines model for Pedido.RateType.
type PedidoRateType string

// Produto defines model for Produto.
type Produto struct {
	Descricao string `json:"descricao"`

	// Id Example: novobanco:primeiro_banco
	Id         string `json:"id"`
	PorOmissao bool   `json:"por_omissao"`
	Rotulo     string `json:"rotulo"`
}

// RespostaErro defines model for RespostaErro.
type RespostaErro struct {
	Erro      DetalheErro `json:"erro"`
	RequestId *string     `json:"request_id,omitempty"`
}

// Saude defines model for Saude.
type Saude struct {
	// Estado Example: ok
	Estado string `json:"estado"`
}

// Titular defines model for Titular.
type Titular struct {
	// DataNascimento Example: 1990-04-12
	DataNascimento   openapi_types.Date `json:"data_nascimento"`
	RendimentoMensal float64            `json:"rendimento_mensal"`
}

// BancoOcupado defines model for BancoOcupado.
type BancoOcupado = RespostaErro

// PedidoInvalido defines model for PedidoInvalido.
type PedidoInvalido = RespostaErro

// TectoExcedido defines model for TectoExcedido.
type TectoExcedido = RespostaErro

// OfertaDeUmBancoJSONRequestBody defines body for OfertaDeUmBanco for application/json ContentType.
type OfertaDeUmBancoJSONRequestBody = OfertaPedido

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ListarBancos Tudo o que a app precisa para montar o formulário adaptativo.
	// (GET /api/v1/bancos)
	ListarBancos(w http.ResponseWriter, r *http.Request)
	// OfertaDeUmBanco Pergunta a UM banco o preço do crédito descrito. Ao vivo.
	// (POST /api/v1/ofertas/{banco})
	OfertaDeUmBanco(w http.ResponseWriter, r *http.Request, banco string)
	// Saude Sondagem de saúde trivial para a plataforma.
	// (GET /healthz)
	Saude(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// ListarBancos Tudo o que a app precisa para montar o formulário adaptativo.
// (GET /api/v1/bancos)
func (_ Unimplemented) ListarBancos(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// OfertaDeUmBanco Pergunta a UM banco o preço do crédito descrito. Ao vivo.
// (POST /api/v1/ofertas/{banco})
func (_ Unimplemented) OfertaDeUmBanco(w http.ResponseWriter, r *http.Request, banco string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Saude Sondagem de saúde trivial para a plataforma.
// (GET /healthz)
func (_ Unimplemented) Saude(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListarBancos operation middleware
func (siw *ServerInterfaceWrapper) ListarBancos(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListarBancos(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OfertaDeUmBanco operation middleware
func (siw *ServerInterfaceWrapper) OfertaDeUmBanco(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "banco" -------------
	var banco string

	err = runtime.BindStyledParameterWithOptions("simple", "banco", chi.URLParam(r, "banco"), &banco, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "banco", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OfertaDeUmBanco(w, r, banco)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Saude operation middleware
func (siw *ServerInterfaceWrapper) Saude(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Saude(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/bancos", wrapper.ListarBancos)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/ofertas/{banco}", wrapper.OfertaDeUmBanco)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.Saude)
	})

	return r
}