# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: simulador-v2
  version: 1.0.0
  description: >-
    Comparação de ofertas de crédito à habitação, e a série temporal de mercado
    de onde ela sai. Serve JSON e mais nada. A interface é uma app React Native
    noutro repositório.

    ⚠️ Dizia aqui «ao vivo», e era verdade até 2026-07-25. A §1 do
    ARQUITETURA.md inverteu-se nesse dia: os simuladores públicos dos bancos
    correm-se em hora morta, por `simulador varrer`, e uma comparação
    responde-se por consulta a essa série e cálculo local. Nenhum pedido de um
    cliente chega a um banco.
servers:
  - url: /
    description: O próprio serviço; os caminhos são absolutos a partir da raiz.
...
paths:
...
  /api/v1/ofertas/{banco}:
    post:
      operationId: ofertaDeUmBanco
      tags: [v1]
      summary: Pergunta a UM banco o preço do crédito descrito. Ao vivo.
      description: >-
        Vai ao simulador público do banco com os valores que a pessoa
        introduziu e devolve o que ele respondeu. Demora o que o banco demorar,
        com um prazo nosso por cima.


        ⚠️ É POST e não GET porque o pedido leva data de nascimento e
        rendimento: num GET isso viajava na query string, e ficava no histórico
        do browser, nos logs de qualquer proxy pelo caminho e no Referer.


        ⚠️ Um banco que não responde a tempo NÃO é um erro deste pedido: a
        resposta é 200 com a oferta em falha e o código `banco_indisponivel`. É
        o mesmo vocabulário do POST /api/v1/comparacoes.
      parameters:
        - name: banco
          in: path
          required: true
          schema:
            type: string
          description: O id do banco, como vem em GET /api/v1/bancos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfertaPedido"
      responses:
        "200":
          description: A oferta do banco, ou a recusa dele com o código nomeado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Oferta"
        "400":
          $ref: "#/components/responses/PedidoInvalido"
        "404":
          description: O id não é banco nenhum.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespostaErro"
        "429":
          $ref: "#/components/responses/TectoExcedido"
        "503":
          $ref: "#/components/responses/BancoOcupado"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    PedidoInvalido:
      description: Pedido inválido, com o campo nomeado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    TectoExcedido:
      description: >-
        Tecto de pedidos por IP excedido.

        ⚠️ A razão mudou com a inversão da §1, e o tecto não. Existia porque cada
        submissão custava dezenas de segundos de scraping a partir do NOSSO IP
        contra os bancos — hoje custa uma consulta, e essa justificação caiu.
        Fica por outra, ordinária e suficiente: isto é JSON público e um tecto é
        o que impede um cliente avariado de esgotar o pool de ligações dos
        outros. É o mesmo endpoint e um limiar muito mais alto; o que não se faz
        é apagar o tecto por a primeira razão ter deixado de valer.
      headers:
        Retry-After:
          description: Segundos a esperar antes de repetir.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
    BancoOcupado:
      description: >-
        Já vão pedidos nossos a mais em curso contra este banco: o tecto de
        concorrência por banco está cheio (ARQUITETURA.md §7.2). Código
        `banco_ocupado`, e a resposta traz `Retry-After`.

        ⚠️ **Não confundir com a oferta em falha `banco_indisponivel`**, que vem
        com 200. Essa diz que o banco não respondeu; esta diz que quem não tem
        lugar somos nós, e o banco está bem. A distinção é para a app poder
        voltar a pedir este banco daqui a um instante em vez de o riscar da
        lista.
      headers:
        Retry-After:
          description: Segundos a esperar antes de voltar a pedir este banco.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RespostaErro"
...
  schemas:
...
    OfertaPedido:
      type: object
      required: [pedido]
      description: >-
        O pedido para UM banco. Não leva lista de bancos: o banco vem no
        caminho, e os produtos são só os desse banco.
      properties:
        pedido:
          $ref: "#/components/schemas/Pedido"
        produtos:
          type: array
          description: Ids dos produtos escolhidos deste banco.
          items:
            type: string
...
    Pedido:
      type: object
      required:
        [
          valor_imovel,
          montante,
          prazo_anos,
          rate_type,
          titulares,
          finalidade,
          localizacao,
        ]
      properties:
        valor_imovel:
          type: number
          format: double
        montante:
          type: number
          format: double
        prazo_anos:
          type: integer
        rate_type:
          type: string
          enum: [variavel, fixa, mista]
        fixed_period_years:
          type: integer
          description: Anos de período fixo; aplica-se a `fixa` e `mista`.
        euribor_indexante:
          type: string
          enum: ["3m", "6m", "12m"]
        titulares:
          type: array
          items:
            $ref: "#/components/schemas/Titular"
        finalidade:
          type: string
          enum: [propria, secundaria, arrendamento]
        localizacao:
          type: string
          example: continente
        garantia_publica:
          type: boolean
        ja_cliente:
          type: boolean
...
    Oferta:
      type: object
      description: >-
        ⚠️ Quando `aplicado` não está vazio, os números não correspondem ao
        pedido — a app é obrigada a mostrar a nota junto do valor. `erro` só
        aparece quando `sucesso` é falso.
      required: [banco_id, banco_nome, sucesso]
      properties:
        banco_id:
          type: string
        banco_nome:
          type: string
        sucesso:
          type: boolean

        # --- Medidos, ou aritmética exacta sobre medições ---
        tan:
          type: number
          format: double
          description: >-
            Spread do degrau de LTV que contém o rácio pedido, mais a Euribor do
            tenor aplicável — ambos medidos. Na fixa e na mista, a taxa da fase
            fixa é a observação do período que o banco pratica.
        spread:
          type: number
          format: double
          description: >-
            ⚠️ Do INTERVALO de LTV medido que contém o rácio pedido, e não de uma
            banda assumida. Num degrau que não se conseguiu resolver é o lado
            mais CARO do intervalo (Directiva 2014/17/UE, Anexo I, Parte II,
            alínea (d)), e `notas` traz a frase que nomeia os dois lados.
        prestacao_mensal:
          type: number
          format: double
          description: >-
            Amortização francesa sobre a `tan` e o montante pedido. É aritmética
            exacta, não medição — e é por isso que responde ao montante desta
            pessoa, e não ao do cenário de referência.

        # --- Derivados: dependem de hipóteses, e declaram-nas ---
        taeg:
          type: number
          format: double
          description: >-
            ⚠️ DERIVADA, não medida. Ver `pressupostos`, que é obrigatório sempre
            que este campo vem preenchido.

            A TAEG depende dos encargos (comissões, imposto, seguros) e o seguro
            de vida depende de quem pede — e a série de mercado é varrida com um
            titular NEUTRO e fictício, sobre um cenário de referência fixo (§4).
            Não existe, portanto, TAEG medida para esta pessoa. O que se faz é
            atribuir a encargos a diferença entre a TAEG e a TAN observadas no
            ponto de referência, e reamortizar sobre os fluxos deste pedido.

            ⚠️ A repartição desses encargos entre comissões, imposto e seguro NÃO
            está medida: é inferida. Quem lê este número tem de o tratar como
            indicativo, e a app é obrigada a mostrar `pressupostos` junto dele —
            é o que o Anexo I e o Anexo II da MCD mandam fazer a um valor que
            depende de hipóteses declaradas.
        mtic:
          type: number
          format: double
          description: >-
            ⚠️ DERIVADO, pela mesma razão e sobre o mesmo modelo de encargos que
            a `taeg`. Ver `pressupostos`.
        pressupostos:
          type: array
          description: >-
            As hipóteses sob as quais a `taeg` e o `mtic` desta oferta foram
            derivados, em português e legíveis por uma pessoa.

            ⚠️ Lista à parte de `notas`, e não misturada nela, de propósito: as
            duas têm estatutos diferentes. Uma `nota` é um aviso sobre o que
            aconteceu a ESTE pedido — o banco encurtou o prazo, o degrau de LTV
            não estava resolvido. Um pressuposto é uma hipótese de cálculo que a
            MCD obriga a declarar junto do número que dela depende (Anexo I,
            Parte II; Anexo II). Empacotadas na mesma lista, a app ficava sem
            forma de as apresentar como o que são.

            ⚠️ Não vazia sempre que `taeg` ou `mtic` vêm preenchidos. Vazia com um
            deles preenchido é defeito nosso, e não um caso legítimo.
          items:
            type: string

        euribor_indexante:
          type: string
        euribor_valor:
          type: number
          format: double
        fases:
          type: array
          items:
            $ref: "#/components/schemas/Fase"
        produtos_aplicados:
          type: array
          items:
            type: string
        aplicado:
          type: object
          description: O que o banco usou de facto, quando difere do pedido. Vazio = igual ao pedido.
          additionalProperties: true
        notas:
          type: array
          items:
            type: string
        fiabilidade:
          type: string
          enum: [por_confirmar, confirmada, em_duvida]
          description: >-
            O que se sabe sobre a grelha de onde este preço saiu, e não sobre o
            preço em si (KAN-49). A sonda confirma — com ~4 pedidos por banco —
            se a grelha guardada ainda descreve o banco.

            ⚠️ `em_duvida` quer dizer que a última sondagem DISCORDOU da grelha
            e que ainda não houve varrimento que resolvesse a discordância. A
            app é **obrigada** a mostrar a nota que vem em `notas`, junto do
            número — é a mesma regra do `aplicado`.

            ⚠️ `confirmada` e `por_confirmar` mostram-se com **silêncio**. Uma
            marca de «confirmada» em toda a gente é ruído com aspecto de
            informação, e treina quem lê a saltar a única que importa.

            ⚠️ **Ausente vale `por_confirmar`**, e nunca `confirmada`: uma
            omissão que valesse «confirmada» afirmava sobre a série inteira uma
            coisa que ninguém mediu. Este servidor manda-o sempre; o campo fica
            opcional porque `required` só tem o que uma oferta não pode não ter.

            ⚠️ E **não leva `default:`** de propósito. Levava, e o
            `openapi-typescript` traduz um campo com omissão para um campo
            OBRIGATÓRIO no tipo gerado — a app passava a afirmar que a resposta
            traz sempre este campo, que é precisamente o que o `required` diz que
            não. A regra fica na descrição, onde não mente ao gerador.
        # ⚠️ O `em_cache` VOLTA (2026-08-07), depois de ter saído com a inversão
        # da §1 e pela razão que essa inversão criou: com tudo a vir da série
        # varrida, o campo era sempre `true` e não informava. Ao vivo volta a
        # separar duas coisas que existem outra vez — uma resposta que acabou de
        # vir do banco, e uma que se serviu de Postgres sem lhe perguntar.
        em_cache:
          type: boolean
          description: >-
            Verdadeiro quando esta oferta se serviu da cache (§7.6) em vez de se
            perguntar ao banco agora.

            ⚠️ **Não é o mesmo que `capturado_em`, e nenhum dos dois substitui o
            outro.** O `capturado_em` diz de QUANDO é o preço, e num acerto de
            cache é o instante em que se falou com o banco — nunca o de agora.
            Este diz se ESTE pedido chegou a sair para o banco. Uma oferta fresca
            e um acerto de um segundo atrás têm `capturado_em` quase igual e
            `em_cache` diferente.

            ⚠️ Ausente vale `false`, e não leva `default:` de propósito — pela
            mesma razão do `fiabilidade` acima: o `openapi-typescript` traduz um
            campo com omissão para um campo OBRIGATÓRIO no tipo gerado.
        capturado_em:
          type: string
          format: date-time
          description: >-
            Quando o preço foi MEDIDO no banco, durante o varrimento — não quando
            esta resposta foi calculada (isso é o `calculado_em` da Comparacao).

            ⚠️ Presente sempre que `sucesso` é verdadeiro, e é o campo mais
            importante desta lista para a honestidade do produto: é o único que
            diz à pessoa que está a ver um preço de ontem à noite e não de agora.
            A app é obrigada a mostrá-lo. Escondê-lo apresentaria dados varridos
            como se fossem uma consulta ao vivo ao banco, que é exactamente o que
            este serviço não faz.
        erro:
          $ref: "#/components/schemas/OfertaErro"
    Fase:
      type: object
      required: [ate_mes, taxa, prestacao]
      properties:
        ate_mes:
          type: integer
        taxa:
          type: number
          format: double
        prestacao:
          type: number
          format: double
...
    RespostaErro:
      type: object
      required: [erro]
      properties:
        erro:
          $ref: "#/components/schemas/DetalheErro"
        request_id:
          type: string
...